Rework input validation

+++ addons/yat/src/commands/ToggleAudio.cs
@@ -12,7 +12,7 @@ public sealed class ToggleAudio : ICommand
 {
 	public CommandResult Execute(CommandData data)
 	{
-		var id = (int)data.Options["-id"];
+		var id = (int)(float)data.Options["-id"];
 		var name = (string)data.Options["-name"];
 
 		if (id == -1 && string.IsNullOrEmpty(name))
